+++ common/src/main/java/net/mehvahdjukaar/jeed/common/ScreenExtensionsHandler.java
@@ -9,7 +9,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.item.TooltipFlag;
 
-import javax.annotation.Nullable;
 import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
@@ -40,19 +39,11 @@ public class ScreenExtensionsHandler {
     }
 
 
-    @ExpectPlatform
-    public static void init(){
-        throw new AssertionError();
-    }
-
-
-    public static void renderEffectTooltip(@Nullable MobEffectInstance effect, Screen screen, PoseStack matrixStack, int x, int y) {
-        if (effect != null) {
-            TooltipFlag flag = Minecraft.getInstance().options.advancedItemTooltips ? TooltipFlag.Default.ADVANCED : TooltipFlag.Default.NORMAL;
-            List<Component> tooltip = EffectRenderer.getTooltipsWithDescription(effect, flag, false);
-            if (!tooltip.isEmpty()) {
-                screen.renderComponentTooltip(matrixStack, tooltip, x, y);
-            }
+    public static void renderEffectTooltip(MobEffectInstance effect, Screen screen, PoseStack matrixStack, int x, int y) {
+        TooltipFlag flag = Minecraft.getInstance().options.advancedItemTooltips ? TooltipFlag.Default.ADVANCED : TooltipFlag.Default.NORMAL;
+        List<Component> tooltip = EffectRenderer.getTooltipsWithDescription(effect, flag, false);
+        if (!tooltip.isEmpty()) {
+            screen.renderComponentTooltip(matrixStack, tooltip, x, y);
         }
     }
 }