+++ forge/src/main/java/net/mehvahdjukaar/jeed/compat/forge/StylishEffectsCompat.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.jeed.compat.forge;
 
+import com.mojang.blaze3d.platform.InputConstants;
 import fuzs.stylisheffects.api.client.MobEffectWidgetContext;
 import fuzs.stylisheffects.api.client.StylishEffectsClientApi;
 import fuzs.stylisheffects.api.client.event.MobEffectWidgetEvent;
@@ -7,6 +8,7 @@ import net.mehvahdjukaar.jeed.Jeed;
 import net.mehvahdjukaar.jeed.api.IEffectScreenExtension;
 import net.mehvahdjukaar.jeed.api.JeedAPI;
 import net.mehvahdjukaar.jeed.common.EffectRenderer;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
 import net.minecraft.client.gui.screens.inventory.EffectRenderingInventoryScreen;
 import net.minecraft.client.gui.screens.inventory.InventoryScreen;
@@ -25,8 +27,11 @@ public class StylishEffectsCompat<T extends EffectRenderingInventoryScreen<?>> i
         });
         MinecraftForge.EVENT_BUS.addListener((final MobEffectWidgetEvent.EffectTooltip evt) -> {
             List<Component> lines = evt.getTooltipLines();
-            //lines.clear();
+            if(!evt.getTooltipFlag().isAdvanced() && !InputConstants.isKeyDown(Minecraft.getInstance().getWindow().getWindow(), Minecraft.getInstance().options.keyShift.getKey().getValue())){
+                lines.remove(lines.size()-1);
+            }
             List<Component> newTooltip = EffectRenderer.getTooltipsWithDescription(evt.getContext().effectInstance(), evt.getTooltipFlag(), false);
+            newTooltip.remove(0);
             lines.addAll(newTooltip);
         });
 
@@ -36,8 +41,8 @@ public class StylishEffectsCompat<T extends EffectRenderingInventoryScreen<?>> i
     }
 
     @Override
-    public MobEffectInstance getEffectAtPosition(T screen, double mouseX, double mouseY, boolean isTooltip) {
-        if (isTooltip) return null;
+    public MobEffectInstance getEffectAtPosition(T screen, double mouseX, double mouseY, CallReason reason) {
+        if (reason != CallReason.RECIPE_KEY) return null; //rest is handled by Stylish Effect events above
         return StylishEffectsClientApi.getEffectScreenHandler().getInventoryHoveredEffect(screen, mouseX, mouseY)
                 .map(MobEffectWidgetContext::effectInstance)
                 .orElse(null);